requirements.txt
pymongo[srv]==3.11
flask
python-dotenv
passlib
flask-jwt-extended